Create SwiftUIViewWrapper.swift
Create a new class that conforms to the UIViewRepresentable protocol, which requires two methods to be implemented: makeUIView(context:) and updateUIView(_:context:).
In the makeUIView(context:) method, you can return an empty UIView, as we will replace this view in the updateUIView(_:context:) method.
In the updateUIView(_:context:) method, you can create an instance of your SwiftUI view, create a UIHostingController and set its rootView to be the SwiftUI view instance. You can then add the UIHostingController's view as a subview of the UIView that was passed in, and finally, call didMove(toParent:) to inform the UIHostingController that it has been added to the parent view controller.
Finally, you can add the SwiftUIViewWrapper to your UIKit view hierarchy as a subview.

In this example, we are simply creating an instance of the SwiftUIViewWrapper, adding it as a subview of our UIViewControllers view, and setting its frame to match the size of the parent view. The SwiftUIViewWrapper will then update itself with the SwiftUIView when necessary.
Overall, embedding a SwiftUI view inside a UIHostingController or using a UIViewRepresentable protocol to create a UIView wrapper are both valid ways to use SwiftUI views inside a UIKit app.

//OMG